+++ src/hellotest/HelloApp.java
@@ -11,9 +11,9 @@ import org.slf4j.LoggerFactory;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import us.kbase.auth2.service.LoggingFilter;
+import us.kbase.auth2.service.exceptions.handler.ExceptionHandler;
 import us.kbase.auth2.service.template.TemplateProcessor;
 import us.kbase.auth2.service.template.mustache.MustacheProcessor;
-import us.kbase.auth2.service.ExceptionHandler;
 
 //TODO WAIT accept json in text/plain and application/x-www-form-urlencoded or manually handle it
