+++ src/us/kbase/auth2/service/exceptions/handler/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package us.kbase.auth2.service;
+package us.kbase.auth2.service.exceptions.handler;
 
 import java.util.HashMap;
 import java.util.List;
